requirements_dev.txt
-r requirements.txt
yapf==0.30.0
pylint==2.5.3
isort==4.3.21
pylint-django==2.3.0
django-extensions==3.0.9
